@@ -1,4 +1,3 @@
-
 {
   "name": "react-static-example-typescript",
   "private": true,
@@ -8,38 +7,46 @@
     "build": "react-static build",
     "bundle": "react-static bundle",
     "export": "react-static export",
-    "serve": "serve dist -p 3000 -s"
+    "serve": "serve dist -p 3000 -s",
+    "lint": "eslint src --ext .js,.jsx,.ts,.tsx",
+    "lint:fix": "yarn lint --fix"
   },
   "dependencies": {
-    "@reach/router": "^1.2.1",
-    "@riboseinc/aperis-doc-pages": "^1.0.0-pre8",
-    "@riboseinc/react-static-plugin-aperis-doc-pages": "^1.0.7",
-    "axios": "^0.19.0",
-    "chroma-js": "^2.1.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-static": "^7.2.0",
-    "react-static-plugin-reach-router": "^7.2.0",
-    "react-static-plugin-sitemap": "^7.2.0",
-    "react-static-plugin-source-filesystem": "^7.2.0",
-    "react-static-plugin-styled-components": "^7.3.0",
-    "react-static-plugin-typescript": "^7.2.0",
-    "styled-components": "^5.1.1"
+    "@reach/router": "^1.3.4",
+    "@riboseinc/aperis-doc-pages": "^1.0.0-pre9",
+    "@riboseinc/react-static-plugin-aperis-doc-pages": "^1.0.25",
+    "axios": "^1.6.2",
+    "chroma-js": "^2.4.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-static": "^7.6.2",
+    "react-static-plugin-reach-router": "^7.6.2",
+    "react-static-plugin-sitemap": "^7.6.2",
+    "react-static-plugin-source-filesystem": "^7.6.2",
+    "react-static-plugin-styled-components": "^7.6.2",
+    "react-static-plugin-typescript": "^7.6.2",
+    "styled-components": "^6.1.1"
   },
   "devDependencies": {
-    "@types/chroma-js": "^2.0.0",
-    "@types/node": "^12.7.2",
-    "@types/reach__router": "^1.2",
-    "@types/react": "^16.9.1",
-    "@types/react-dom": "^16.8.5",
-    "@types/react-helmet": "^6.1.0",
+    "@types/chroma-js": "^2.4.3",
+    "@types/node": "^20.9.4",
+    "@types/reach__router": "^1.3",
+    "@types/react": "^18.2.38",
+    "@types/react-dom": "^18.2.17",
+    "@types/react-helmet": "^6.1.9",
     "@types/react-hot-loader": "^4.1.0",
-    "@types/styled-components": "^5.1.2",
-    "@types/webpack-env": "^1.14.0",
-    "react-hot-loader": "^4.12.11",
+    "@types/styled-components": "^5.1.32",
+    "@types/webpack-env": "^1.18.4",
+    "@typescript-eslint/eslint-plugin": "^6.12.0",
+    "@typescript-eslint/parser": "^6.12.0",
+    "eslint": "^8.54.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "react-hot-loader": "^4.13.1",
     "react-static-plugin-file-watch-reload": "^1.0.4",
-    "serve": "^11.1.0",
-    "typescript": "^3.9.6"
-  }
+    "serve": "^14.2.1",
+    "typescript": "^5.3.2"
+  },
+  "packageManager": "yarn@1.22.10"
 }
-      
